// backend/src/security/role-type.ts
'use strict';

export enum RoleType {
  USER = 'user',
  ADMIN = 'admin',
  ANONYMOUS = 'anonymous',
  MANAGER = 'manager',
  REPORTER = 'reporter',
  VIEW_PERSONNEL_INFO = 'view_personnel_info',
  PERSONNEL_REVIEW = 'personnel_review',
  VIEW_HR = 'view_hr',
  VIEW_STATISTIC = 'view_statistic',
  RECRUITMENT_MANAGEMENT = 'recruitment_management',
  RECRUITMENT_MANAGER = 'recruitment_manager',
}